feat: 优化微信登录流程和订阅权限管理 - 登录时只显示加载弹窗，成功直接跳转；权限请求移到保存时；微信用户无需重复授权；新增用户偏好设置；修复JWT配置问题；优化UI交互体验

+++ reminder-backend/reminder-core/src/main/java/com/core/reminder/security/JwtTokenProvider.java
@@ -23,7 +23,7 @@ public class JwtTokenProvider {
     private String jwtSecret;
 
     @Value("${app.jwtExpirationInMs}")
-    private int jwtExpirationInMs;
+    private long jwtExpirationInMs;
 
     private Key key;
